Replace ship user input correction to views instead of text input. Fixed skill autocorrect

+++ cogs/wows/skill.py
@@ -5,6 +5,7 @@ from discord import app_commands, Embed, Interaction
 from discord.ext import commands
 
 from mackbot.exceptions import NoSkillFound, SkillTreeInvalid
+from mackbot.utilities.discord.views import ConfirmCorrectionView
 from mackbot.utilities.correct_user_mispell import correct_user_misspell
 from mackbot.utilities.discord.items_autocomplete import auto_complete_skill_name
 from mackbot.utilities.find_close_match_item import find_close_match_item
@@ -24,6 +25,7 @@ class Skill(commands.Cog):
 	@app_commands.autocomplete(skill_name=auto_complete_skill_name)
 	async def skill(self, interaction: Interaction, skill_name: str):
 		# get information on requested skill
+
 		try:
 			embed = Embed(description=f"## {skill_name.title()}")
 			skill_data = get_skill_data(skill_name)
@@ -47,7 +49,7 @@ class Skill(commands.Cog):
 			if len(skill_data) > 1:
 				embed.description = f"## {name}\n May refer to a skill in one of these trees."
 
-			await interaction.response.send_message(embed=embed, file=skill_image)
+			await interaction.channel.send(embed=embed, file=skill_image)
 
 		except Exception as e:
 			logger.info(f"Exception in skill {type(e)}: {e}. No skill found for {skill_name}")
@@ -58,12 +60,27 @@ class Skill(commands.Cog):
 				embed = Embed(title=f"Skill {skill_name} is not understood.\n", description="")
 				if len(closest_match) > 0:
 					embed.description += f'\nDid you mean **{closest_match[0]}**?'
-					embed.description += "\n\nType \"y\" or \"yes\" to confirm."
 					embed.set_footer(text="Response expires in 10 seconds")
-				msg = await interaction.channel.send(embed=embed)
-				await correct_user_misspell(self.bot, interaction, Skill, "skill", closest_match[0])
-				await msg.delete()
+				confirm_view = ConfirmCorrectionView(
+					closest_match,
+				)
+
+				new_line_prefix = "\n-# - "
+				embed.description = f'\nDid you mean **{closest_match}**?\n-# Other possible matches are: {new_line_prefix}{new_line_prefix.join(i.title() for i in closest_match[1:])}'
+				embed.set_footer(text="Response expire in 10 seconds")
+				msg = await interaction.channel.send(
+					embed=embed,
+					view=confirm_view,
+					delete_after=10
+				)
+				await confirm_view.wait()
+
+				if confirm_view.value:
+					await correct_user_misspell(self.bot, interaction, Skill, "skill", closest_match[0])
+				else:
+					await msg.delete()
+
 			if type(e) == SkillTreeInvalid:
 				embed = Embed(title=f"Skill tree is not understood.\n", description="")
 				embed.description += f'\n{e}'
-				await interaction.response.send_message.send(embed=embed)
+				await interaction.channel.send(embed=embed)
